+++ src/spVIPES/data/_manager.py
@@ -15,12 +15,12 @@ import scvi
 from mudata import MuData
 from rich import box
 from rich.console import Console
-from scvi._types import AnnOrMuData
 from scvi.data import AnnTorchDataset
 from scvi.utils import attrdict
 from torch.utils.data import Subset
 
 from . import _constants
+from ._types import AnnOrMuData
 from ._utils import (
     _assign_adata_uuid,
     _check_if_view,